+++ ReleaseNotes.md
@@ -8,6 +8,7 @@ Changed default for layout to "js" if built with JavaScript support.
 
 - All configuration options for autoscan directories are available in the web UI now, including scripts and storage in the database.
 - New configuration option for autoscan directories to rescan files that ended up as plain object in the first run.
+- Several script options to control audio layout
 
 ### FIXES